fix yuma123 path substitutions in configure.ac
Generate correct values so these can later be used by makefile templates.
In particular, this replaces the string "NONE" with a meaningful value
in config.status.  For example:
  S["yangdir"]="NONE/share/yuma/modules"

Use the path variables $datarootdir and $includedir defined by autoconf
for compatibility with packaging systems that set these values on the
configure commandline.