use xml macros generated by autoconf
Adjust the automake templates to use XML_LIBS and XML_CPPFLAGS.

In some Makefiles, this also fixes the libxml include paths constructed
with ${includedir} which is the location into which yuma123 installs
headers, not the location of system headers.  Supplying a --prefix=
option to configure would cause the the compiler to look in the wrong
place for libxml headers.

Also touch up the quilt patch used for debian package builds.