commandline args take precendence over some environment variables
The netconfd manpage claims that some commandline arguments override
the equivalent environment variable, however this is not always true.
The NCX_YUMA_HOME_STARTUP_FILE macro, for example, begins with the
string "$YUMA_HOME".  ncx_get_source() will always try to expand from
the environment variable even if the --yuma-home option was provided.

Additionally, a environment variable that is not set will expand to an
empty string.  This can happen if the --yuma-home argument is supplied on
the commandline; in this case a home directory was supplied and should
be checked for a startup configuration (yumahome is not NULL) but since
$YUMA_HOME is expanded anyways the configuration file is never found,
even if it exists.

Since commandline parameters are processed very early, check to see if
there is already a value available for files/paths influenced by the
YUMA_HOME, YUMA_MODPATH, YUMA_DATAPATH and YUMA_RUNPATH environment
variables and, if so, use that cached value instead.  This will allow
values given on the commandline "override" the environment as the order
in which options are processed will determine priority.