run commit-complete user callbacks after loading startup config
User callbacks defined by the SIL modules are run for each stage
of commit, including the final AGT_CB_COMMIT stage.  However, the
user callbacks to be run after a commit are not called.  After
loading the configuration file and marking the running configuration
ready, run the user commit-complete callbacks.