test: install SIL modules where netconfd will find them
dh_auto_configure specifies a library directory on the configure command
line which causes the yuma libraries to be installed in a directory
below /usr/lib.  This also causes the yuma libraries to look for SIL
modules in that same location.  Update the debian platform validation
script (debian.sh) so that it also installs the test SIL modules in a
location that netconfd can find by default.