return status value from val_make_sprintf_string
Change val_make_sprintf_string() to take a string pointer as a parameter
and return a status_t.  This will allow calling functions to have some
indication of why formatting the string failed and will allow returning
partial results.

In val_get_leafref_targval, also change the type of target_str to xmlChar*
and use xmlStrcmp() to compare this to VAL_STRING(leafref_val), since
that is also of type xmlChar*.