do not abort as a result of invalid data supplied by user
If a netconf client attempts to create a leaf-list configuration
node of a union type with an invalid value, netconfd will generate an
internal error caused by a failure to format the instance ID used in the
rpc-reply's error-path, instead of any indication that the supplied value
was unusable.  Currently, this causes netconfd to raise an assertion
because val_make_sprintf_string() outputs a NULL pointer.

In order to provide feedback to the client about where things went wrong,
omit the leaf-list element and output the node path up to the offending
value.  Preserve the error status and return it to the calling function.

Examine the return value from agt_record_error_errinfo().  If the error
is not catastrophic, use the resulting string in the rpc-reply message.