yangrpc: remove asserts and add cleanup code
Make this library friendlier to applications.  Remove assert()s where
the application can handle the error.  Fill in yangrpc_close() so that
an old, possibly failed, connection can be discarded and another created.

One assert remains in a function that has no return value and serves as
a callback.  Removing this assert will have to be part of a cleanup in
other libraries.