support leaf-list default values
Yang 1.1 allows a leaf-list to have one or more default values.
Parse default substatements to a leaf-list.  Default values are only in
use if the leaf-list is not present in the data tree.  For this reason,
the usual method of applying default values to items in the data tree
does not work.  Instead, when retrieving data, examine the schema tree
for objects that have leaf-list(s) as children and supply values when
a child leaf-list is missing from the data tree.

Signed-off-by: Vladimir Vassilev <vladimir@lightside-instruments.com>