fix double-free when deviation replaces type in augment
When an augment statement is expanded, the objects defined by the augment
statement block are cloned to fill in the schema tree.  A deviation that
replaces a type frees the typ_def_t pointer in that object's definition
and replaces it with the type referenced by the deviation.  Since the
original typ_def_t pointer can also be found in some node's augment
datadef queue, it will be freed a second time during shutdown.

In apply_object_deviate_type(), only free the typdef pointer if the
object is not a clone.